Fix broken assumption where stat.st_nlink == 1
Normally, any directory has a st_nlink of at least 2 (. and ..).
On some systems, a directory's st_nlink is set to 1 where it is too
expensive to compute the correct value (e.g. remote filesystems).
Instead of treating this as a legitimate value (i.e. that the directory
contains nothing to link from), special case this and force a complete
scan of the directory's contents.

Further explanation:
http://cygwin.com/ml/cygwin/2008-06/msg00056.html
http://cygwin.com/ml/cygwin-apps/2008-08/msg00264.html

Signed-off-by: Yaakov Selkowitz <yselkowitz@users.sourceforge.net>